// src/types/api.ts
export interface components {
    schemas: {
// ...
        User: {
            /** @example 12 */
            id?: number;
            /**
             * Format: email
             * @example ratna@example.com
             */
            email?: string;
            /** @example Ratna Dewi */
            name?: string;
            /**
             * @example female
             * @enum {string}
             */
            gender?: "male" | "female";
            /**
             * Format: date
             * @example 1985-06-12
             */
            birthdate?: string;
            /** @example +6281234567890 */
            phone?: string;
            /** @example 17 */
            province_id?: number;
            /** @example 217 */
            kabkota_id?: number;
            /** Format: date-time */
            email_verified_at?: string | null;
            children?: components["schemas"]["UserChild"][];
            /** Format: date-time */
            created_at?: string;
            /** Format: date-time */
            updated_at?: string;
        };
        ChildInput: {
            /**
             * @example male
             * @enum {string}
             */
            gender: "male" | "female";
            /**
             * Format: date
             * @example 2014-09-21
             */
            birthdate: string;
        };
        UserChild: components["schemas"]["ChildInput"] & {
            /** @example 41 */
            id?: number;
            /** @example 12 */
            user_id?: number;
            /** Format: date-time */
            created_at?: string;
        };
// ...
    };
// ...
}